Update patch set 1
Patch Set 1:

If there are no scripts in the wild the depend on flashing bootloader / radio image from update.zip this is fine.

Otherwise we might should consider if we should add a field to the fstab if the partition should be formatable to avoid noobs from killing their bootloader.

Patch-set: 1