Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

No, I want to keep this in here and hard code those in the recovery as "DO NOT FUCKING TOUCH THIS SHIT". I suspect some OTA update may use these things, and don't want to break that.

Jared, please adjust recovery code and submit a fix. This is rather low priority for me... until I need to release a new recovery that this would be a problem :)

Patch-set: 1