Update patch set 1
Patch Set 1:

Where is your source repository? I don't think Koush has done any kernel tuning on the NS (uses stock kernel? slacker!), and we definitely don't have a repository for it in CM. I'm thinking that this should probably be coordinated/maintained properly if we are switching to something custom.

Patch-set: 1