Update patch set 3
Patch Set 3:

After some clarification in IRC, I understand that this is not the correct approach. I'm working on that.

However, I would like to make sure I understand something correctly. When you say that this will make every other device build these, how is that? It seems to me that it checks whether the current target is Crespo, and only proceeds if that is true. As long as the target isn't crespo, it should skip the rest of the file.
Am I incorrect?

Patch-set: 3