Update patch set 1
Patch Set 1:

The comment should explain your intention and why it is needed. As far as I can tell it is: "Disable software GLES renderer on crespo". Except we already had the hardware renderer enabled. It also should mention which applications are actually affected so we can verify they are fixed. At the moment I can't attempt to verify anything except maybe "the phone seems to work" (or not as the case may be).

Please, write your commit comments with the aim of explaining your changeset to someone else. Even yourself 6 months later as you'll have forgotten why you removed that line without a suitable explanation. Not that this is just you - most of the cyanogenmod commits have piss poor commit messages. Everyone should look over the git commits or the linux kernel commits as a guide.

Patch-set: 1