Update patch set 2
Patch Set 2:

Nebojsa Cvetkovic:

Take a look more carefully. commit change has being changed. On the other hand I always review my ouwn code since I decide when in my opinion is resdy to be released or not. I cand send to gerrit and no score or set -1 if I think is not fully tested or is not still the time to commit that change

Patch-set: 2