Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

A better thing to do would be to figure out why Android is choosing the software renderer in the first place. If it's working on stock, then it's likely that all the hacks we've added to CM to support other devices and old graphics drivers has broken something.

Patch-set: 2