Update patch set 1
Patch Set 1:

Reading Dianne's post on G+, it seems that they changed the way hardware accelerations works to suite Nexus S: they reverted to sw rendering for some parts of the UI in order to avoid excessive usage of RAM and give a smoother experience.

With this change, the experience is better than before but not perfect.

Patch-set: 1