Update patch set 2
Patch Set 2:

@Kalim -

Tested with Crespo4g.  I built with this and had several freezes.  I didn't keep it long enough to get logcats, needed a functioning phone as I was traveling.  Will try and build again possibly tomorrow and see if the issue continues to capture a logcat.  Do you have a new herring_defconfig so that the kernel can be built from source?

Patch-set: 2