Update patch set 2
Patch Set 2:

Ok patrick i will remove the dot tomorrow. About x-proc: i have a 9023 so no chance to test it! But i know the meaning from some photographer's and they told me that x-processing is a common used word! so if it's ok for you i prefer to leave it like that.

Patch-set: 2