Update patch set 1
Patch Set 1:

So, this is supposed to be on the cm-10.2 branch, but that doesn't exist yet, and I couldn't find a way to make it with gerrit. 

Obviously this should not be pulled into the cm-10.1 branch...

How would I go about fixing that?

Patch-set: 1
Reviewer: Joel Teichroeb <5811@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0