Update patch set 1
Patch Set 1:

I guess my only concern is that users are used to the current kernel interface for configuring the backlight/backlight delay, but with these changes it will take that away. Right with these changes it is impossible to not have the lights turn off after 5 seconds of inactivity. I think it might make sense to just wait until the power manager button backlight code is merged so there is no loss in functionality.

Maybe I'm being too idealistic here though as cm-10.2 is unstable still.

Patch-set: 1