Update patch set 1
Patch Set 1: Code-Review-1

I assume you were looking at this.
http://source.android.com/devices/low-ram.html


It says "For the really low-memory devices, we recommend the JIT be disabled entirely."

I don't think 512MB qualifies as "really low-memory". It says it will only save 3-6MB of RAM. I doubt that is worth it for the performance hit.

I think that option is more meant for phones with less than 512MB of RAM.

Patch-set: 1
Label: Code-Review=-1