Update patch set 1
Patch Set 1:

@Taichi

Ever since that commit I've seem more people complain about things running really slowly when a larger number of apps are open. I'm not certain it is the commit I reverted that was causing the problems, but I wanted to see if the problems would go away with it reverted.

Have you seen any performance problems recently?

Patch-set: 1