Update patch set 1
Patch Set 1:

I'm not sure why this would benefit crespo. From my understanding of the feature, it is supposed to make it so that we do not clean the screen after the bootloader is done. This doesn't really matter on crespo, or on most devices, as the CM boot animation starts with a blank frame. My guess is that the feature is intended for something like the One+ One where they can make it so that the bootloader and android boot animation mesh nicely. 

With that said, I'm building with this change now to see if I see any difference.

Patch-set: 1