Update patch set 1
Patch Set 1: Code-Review-1

I guess I misunderstood it a bit. It makes it so that the splash keeps going while the interface is coming up. Personally, I think it looks worse. Before this patch it instantly went from the splash screen to the interface. Now it weirdly transitions.

Patch-set: 1
Label: Code-Review=-1